fix app id getter

+++ App/params.py
@@ -17,5 +17,5 @@ class AppParams(metaclass=Params):
     scopes = ListValidator('scopes', '应用权限列表').element(Validator().to(Scope.get_by_name))
     premises = ListValidator('premises', '应用要求列表').element(Validator().to(Premise.get_by_name))
 
-    app = Validator('app_id', '应用ID', 'app').to(int).to(App.get_by_id)
+    app = Validator('app_id', '应用ID', 'app').to(App.get_by_id)
     user_app = Validator('user_app_id', '用户绑定应用ID', 'user_app').to(UserApp.get_by_id)
